+++ src/preview/browser/builder.rs
@@ -1,27 +1,8 @@
-// FontFor: find fonts which can show a specified character
-// Copyright (C) 2019 7sDream <i@7sdre.am> and contributors
-//
-// This file is part of FontFor.
-//
-// FontFor is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-mod static_page_server;
-
-use static_page_server::StaticPageServer;
-
-use crate::font::{Family, GetValueByLang};
-use std::iter::FromIterator;
+use {
+    super::SingleThreadServer,
+    crate::font::{Family, GetValueByLang},
+    std::iter::FromIterator,
+};
 
 pub struct Builder<'a> {
     families: Vec<&'a str>,
@@ -69,7 +50,7 @@ impl<'a> Builder<'a> {
         )
     }
 
-    pub fn build_for(self, c: char) -> StaticPageServer {
-        StaticPageServer::new(self.build_html(c))
+    pub fn build_for(self, c: char) -> SingleThreadServer {
+        SingleThreadServer::new(self.build_html(c))
     }
 }